scenarios/deck_builder/pyproject.toml
[project]
name = "deck-builder"
version = "1.0.0"
description = "PowerPoint deck builder that converts markdown to PPTX"
requires-python = ">=3.11"
dependencies = [
    "python-pptx>=1.0.0",
    "click>=8.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff",
    "pyright",
]

[project.scripts]
deck-builder = "scenarios.deck_builder.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "standard"